Made major modifications to file parsing
After a lot of painful failures, I'm not sure that the kind of
polymorphism I was trying to implement is actually possible in rust, so
I totally rewrote all the fasta/fastq parsing code using the seq_io
library. It works now, and allows the user to do the main thing I was
trying to accomplish all along: pass a file path to a method that could
be fasta or fastq, gzipped or uncompressed, and return something that
can iterate over its entries without worrying about what kind of file it
is. Hooray!