Fixed missing alignment score bug
Some aligners do not report an alignment score for unmapped reads, which
was causing the program to exit with an error. I changed it so that
missing alignment scores do not cause an error if the associated read is
unmapped, but instead are equivalent to a very negative alignment score.